Add util/bulk-fs-task.js
