+++ lib/md5-cache.js
@@ -6,34 +6,7 @@ var lodash = require('lodash');
 var pluginCompat = require('./util/plugin-compat');
 var promisify = require('./util/promisify');
 var values = require('./util/Object.values');
-
-var bulkFsTask = function(array, each) {
-  return new Promise(function(resolve, reject) {
-    var ops = 0;
-    var out = [];
-    array.forEach(function(item, i) {
-      out[i] = each(item, function(back, callback) {
-        ops++;
-        return function(err, value) {
-          try {
-            out[i] = back(err, value, out[i]);
-          }
-          catch (e) {
-            return reject(e);
-          }
-
-          ops--;
-          if (ops === 0) {
-            resolve(out);
-          }
-        };
-      });
-    });
-    if (ops === 0) {
-      resolve(out);
-    }
-  });
-};
+var bulkFsTask = require('./util/bulk-fs-task');
 
 class Md5Cache {
   apply(compiler) {
